+++ lib_xcore_math/src/scalar/scalar_float_s64.c
@@ -12,18 +12,7 @@
 float_s32_t float_s64_to_float_s32(
     const float_s64_t x)
 {
-  float_s32_t res = {
-    .exp = x.exp,
-    .mant = (int32_t) x.mant
-  };
-
-  headroom_t x_hr = HR_S64(x.mant);
-
-  if(x_hr < 32) {
-    right_shift_t shr = 32 - x_hr;
-    res.mant = (int32_t) (x.mant >> shr);
-    res.exp = x.exp + shr;
-  }
-
+  float_s32_t res;
+  res.mant = s64_to_s32(&res.exp, x.mant, x.exp);
   return res;
 }