+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ lib_xcore_math change log
 UNRELEASED
 ----------
 
+  * ADDED: `s16_ashr`, `s32_ashr` and `s64_ashr` APIs
+  * ADDED: `f32_to_s32`, `float_s32_use_exponent`, `float_s32_to_s32`
+    and `float_s64_to_s32` APIs
   * REMOVED: `xcommon` build system support
   * CHANGED: Renamed `float_s32_to_float` to `float_s32_to_f32`
   * CHANGED: Renamed `float_s32_to_double` to `float_s32_to_f64`